+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 3.2 alpha 2 - Unreleased
 
 - Implemented `DatabaseOperations.last_executed_query()`.
-- Install now requires snowflake-connector-python 2.7.2 (previously no minimum
+- Install now requires snowflake-connector-python 2.7.4 (previously no minimum
   version was specified).
 - Made `Cursor.execute()` interpolate SQL when `params=()`, consistent with
   other database drivers.